Use page endpoints for fetching data
