+++ package.json
@@ -10,11 +10,12 @@
     "check": "svelte-check --tsconfig ./tsconfig.json",
     "check:watch": "svelte-check --tsconfig ./tsconfig.json --watch",
     "lint": "prettier --ignore-path .gitignore --check --plugin-search-dir=. . && eslint --ignore-path .gitignore .",
-    "format": "prettier --ignore-path .gitignore --write --plugin-search-dir=. ."
+    "format": "prettier --ignore-path .gitignore --write --plugin-search-dir=. .",
+    "sync": "svelte-kit sync"
   },
   "devDependencies": {
     "@sveltejs/adapter-static": "^1.0.0-next.28",
-    "@sveltejs/kit": "1.0.0-next.278",
+    "@sveltejs/kit": "1.0.0-next.290",
     "@types/js-yaml": "^4.0.5",
     "@typescript-eslint/eslint-plugin": "^5.13.0",
     "@typescript-eslint/parser": "^5.13.0",
